# Parallel sort for ConcurrentGrouper