--- a/processing/src/main/java/io/druid/query/groupby/epinephelinae/ConcurrentGrouper.java
+++ b/processing/src/main/java/io/druid/query/groupby/epinephelinae/ConcurrentGrouper.java
@@ -28,17 +28,20 @@ import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.ListeningExecutorService;
 import io.druid.java.util.common.ISE;
 import io.druid.query.AbstractPrioritizedCallable;
+import io.druid.query.QueryInterruptedException;
 import io.druid.query.aggregation.AggregatorFactory;
 import io.druid.query.groupby.orderby.DefaultLimitSpec;
 import io.druid.segment.ColumnSelectorFactory;
 
-import javax.annotation.Nullable;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
+import java.util.concurrent.CancellationException;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
@@ -72,9 +75,10 @@ public class ConcurrentGrouper<KeyType> implements Grouper<KeyType>
   private final DefaultLimitSpec limitSpec;
   private final boolean sortHasNonGroupingFields;
   private final Comparator<Grouper.Entry<KeyType>> keyObjComparator;
-  @Nullable
   private final ListeningExecutorService grouperSorter;
   private final int priority;
+  private final boolean hasQueryTimeout;
+  private final long queryTimeoutAt;
 
   private volatile boolean initialized = false;
 
@@ -91,8 +95,10 @@ public class ConcurrentGrouper<KeyType> implements Grouper<KeyType>
       final int concurrencyHint,
       final DefaultLimitSpec limitSpec,
       final boolean sortHasNonGroupingFields,
-      @Nullable final ListeningExecutorService grouperSorter,
-      final int priority
+      final ListeningExecutorService grouperSorter,
+      final int priority,
+      final boolean hasQueryTimeout,
+      final long queryTimeoutAt
   )
   {
     Preconditions.checkArgument(concurrencyHint > 0, "concurrencyHint > 0");
@@ -120,8 +126,10 @@ public class ConcurrentGrouper<KeyType> implements Grouper<KeyType>
     this.limitSpec = limitSpec;
     this.sortHasNonGroupingFields = sortHasNonGroupingFields;
     this.keyObjComparator = keySerdeFactory.objectComparator(sortHasNonGroupingFields);
-    this.grouperSorter = grouperSorter;
+    this.grouperSorter = Preconditions.checkNotNull(grouperSorter);
     this.priority = priority;
+    this.hasQueryTimeout = hasQueryTimeout;
+    this.queryTimeoutAt = queryTimeoutAt;
   }
 
   @Override
@@ -238,45 +246,50 @@ public class ConcurrentGrouper<KeyType> implements Grouper<KeyType>
 
   private boolean isParallelSortAvailable()
   {
-    return grouperSorter != null && concurrencyHint > 1;
+    return concurrencyHint > 1;
   }
 
   private List<Iterator<Entry<KeyType>>> parallelSortAndGetGroupersIterator()
   {
     // The number of groupers is same with the number of processing threads in grouperSorter
-    final List<ListenableFuture<Iterator<Entry<KeyType>>>> futures = groupers
-        .stream()
-        .map(grouper ->
-                 grouperSorter.submit(
-                     new AbstractPrioritizedCallable<Iterator<Entry<KeyType>>>(priority)
-                     {
-                       @Override
-                       public Iterator<Entry<KeyType>> call() throws Exception
-                       {
-                         synchronized (grouper) {
-                           return grouper.iterator(true);
-                         }
-                       }
-                     }
-                 )
-        ).collect(Collectors.toList());
+    final ListenableFuture<List<Iterator<Entry<KeyType>>>> future = Futures.allAsList(
+        groupers.stream()
+                .map(grouper ->
+                         grouperSorter.submit(
+                             new AbstractPrioritizedCallable<Iterator<Entry<KeyType>>>(priority)
+                             {
+                               @Override
+                               public Iterator<Entry<KeyType>> call() throws Exception
+                               {
+                                 return grouper.iterator(true);
+                               }
+                             }
+                         )
+                )
+                .collect(Collectors.toList())
+    );
 
     try {
-      return Futures.allAsList(futures).get();
+      final long timeout = queryTimeoutAt - System.currentTimeMillis();
+      return hasQueryTimeout ? future.get(timeout, TimeUnit.MILLISECONDS) : future.get();
     }
-    catch (InterruptedException | ExecutionException e) {
-      throw new RuntimeException(e);
+    catch (InterruptedException | TimeoutException e) {
+      future.cancel(true);
+      throw new QueryInterruptedException(e);
+    }
+    catch (CancellationException e) {
+      throw new QueryInterruptedException(e);
+    }
+    catch (ExecutionException e) {
+      throw new RuntimeException(e.getCause());
     }
   }
 
   private List<Iterator<Entry<KeyType>>> getGroupersIterator(boolean sorted)
   {
     return groupers.stream()
-                   .map(grouper -> {
-                     synchronized (grouper) {
-                       return grouper.iterator(sorted);
-                     }
-                   }).collect(Collectors.toList());
+                   .map(grouper -> grouper.iterator(sorted))
+                   .collect(Collectors.toList());
   }
 
   @Override