```ConcurrentGrouper.iterator()``` is one of the bottlenecks for groupBy v2 because it sorts the underlying groupers sequentially to make a merged iterator when ```sorted``` is true. With this patch, historicals can sort the groupers in parallel. The performance gain is about 10% without serde and spilling. Here is the benchmark result.

```
master

Benchmark                                              (defaultStrategy)  (initialBuckets)  (numProcessingThreads)  (numSegments)  (queryGranularity)  (rowsPerSegment)  (schemaAndQuery)  Mode  Cnt       Score       Error  Units
GroupByBenchmark.queryMultiQueryableIndex                             v2                -1                       8             16                 all            100000           basic.A  avgt   30  218997.450 ±  2063.168  us/op
GroupByBenchmark.queryMultiQueryableIndexWithSerde                    v2                -1                       8             16                 all            100000           basic.A  avgt   30  339409.220 ± 22152.255  us/op
GroupByBenchmark.queryMultiQueryableIndexWithSpilling                 v2                -1                       8             16                 all            100000           basic.A  avgt   30  782910.486 ±  4979.416  us/op


patch

Benchmark                                              (defaultStrategy)  (initialBuckets)  (numProcessingThreads)  (numSegments)  (queryGranularity)  (rowsPerSegment)  (schemaAndQuery)  Mode  Cnt       Score      Error  Units
GroupByBenchmark.queryMultiQueryableIndex                             v2                -1                       8             16                 all            100000           basic.A  avgt   30  197634.705 ± 1401.627  us/op
GroupByBenchmark.queryMultiQueryableIndexWithSerde                    v2                -1                       8             16                 all            100000           basic.A  avgt   30  321135.049 ± 2654.269  us/op
GroupByBenchmark.queryMultiQueryableIndexWithSpilling                 v2                -1                       8             16                 all            100000           basic.A  avgt   30  779848.190 ± 5956.896  us/op
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/druid-io/druid/4660)
<!-- Reviewable:end -->