@jihoonson 2017-08-09
@leventov @gianm thanks for the review. Addressed your comments.